Rename the hardware interface package
from vesc_hi to vesc_hw_interface
