+++ vesc_hw_interface/src/vesc_servo_controller.cpp
@@ -16,7 +16,7 @@
 *
 *********************************************************************/
 
-#include "vesc_hi/vesc_servo_controller.h"
+#include "vesc_hw_interface/vesc_servo_controller.h"
 
 namespace vesc_hi {
