+++ vesc_hw_interface/vesc_hw_interface_plugin.xml
@@ -1,4 +1,4 @@
-<library path="lib/libvesc_hi">
+<library path="lib/libvesc_hw_interface">
   <class name="vesc_hi/VescHI" type="vesc_hi::VescHI" base_class_type="hardware_interface::RobotHW">
     <description>
         plugin of VESC hardware interface
